requirements.txt
openpyxl
python-barcord
python-barcode[images]
fpdf
pyinstaller
